+++ src/health-models/azext_health_models/aaz/latest/monitor/health_models/authentication_setting/_show.py
@@ -13,6 +13,7 @@ from azure.cli.core.aaz import *
 
 @register_command(
     "monitor health-models authentication-setting show",
+    is_preview=True,
 )
 class Show(AAZCommand):
     """Get a AuthenticationSetting