+++ src/health-models/azext_health_models/aaz/latest/monitor/health_models/identity/_assign.py
@@ -13,6 +13,7 @@ from azure.cli.core.aaz import *
 
 @register_command(
     "monitor health-models identity assign",
+    is_preview=True,
 )
 class Assign(AAZCommand):
     """Assign the user or system managed identities.