+++ src/horizondb/azext_horizondb/_client_factory.py
@@ -40,3 +40,7 @@ def cf_horizondb_clusters(cli_ctx, _):
 
 def cf_horizondb_parameter_groups(cli_ctx, _):
     return get_horizondb_management_client(cli_ctx).horizon_db_parameter_groups
+
+
+def cf_horizondb_firewall_rules(cli_ctx, _):
+    return get_horizondb_management_client(cli_ctx).horizon_db_firewall_rules