+++ src/networkcloud/azext_networkcloud/aaz/latest/networkcloud/baremetalmachine/_replace.py
@@ -23,9 +23,9 @@ class Replace(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2026-07-01",
+        "version": "2026-05-01-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/baremetalmachines/{}/replace", "2026-07-01"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/baremetalmachines/{}/replace", "2026-05-01-preview"],
         ]
     }
 
@@ -218,7 +218,7 @@ class Replace(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2026-07-01",
+                    "api-version", "2026-05-01-preview",
                     required=True,
                 ),
             }