+++ src/networkcloud/azext_networkcloud/aaz/latest/networkcloud/cluster/_create.py
@@ -29,9 +29,9 @@ class Create(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2026-07-01",
+        "version": "2026-05-01-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/clusters/{}", "2026-07-01"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/clusters/{}", "2026-05-01-preview"],
         ]
     }
 
@@ -269,9 +269,6 @@ class Create(AAZCommand):
             options=["password"],
             help="The password of the service principal.",
             required=True,
-            blank=AAZPromptPasswordInput(
-                msg="Password:",
-            ),
         )
         cluster_service_principal.principal_id = AAZStrArg(
             options=["principal-id"],
@@ -711,7 +708,7 @@ class Create(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2026-07-01",
+                    "api-version", "2026-05-01-preview",
                     required=True,
                 ),
             }